requirements.txt
<<<<<<< HEAD
amqp==2.4.1
appnope==0.1.0
argon2-cffi==19.1.0
backcall==0.1.0
billiard==3.5.0.5
biopython==1.72
bleach==3.0.2
boto3==1.9.102
botocore==1.12.102
celery==4.2.1
certifi==2018.11.29
cffi==1.12.0
confusable-homoglyphs==3.2.0
cycler==0.10.0
Cython==0.29.1
decorator==4.3.0
defusedxml==0.5.0
Django==2.1.7
django-crispy-forms==1.7.2
django-jquery==3.1.0
django-mysql==2.4.1
django-nocaptcha-recaptcha==0.0.20
django-recaptcha==2.0.2
django-registration==3.0
docutils==0.14
entrypoints==0.2.3
ipykernel==5.1.0
ipython==7.1.1
ipython-genutils==0.2.0
ipython-sql==0.3.9
ipywidgets==7.4.2
jedi==0.13.1
Jinja2==2.10
jmespath==0.9.4
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
kiwisolver==1.0.1
kombu==4.3.0
llvmlite==0.26.0
MarkupSafe==1.0
matplotlib==2.2.3
mistune==0.8.4
mysqlclient==1.4.2.post1
nbconvert==5.4.0
nbformat==4.4.0
notebook==5.7.0
numba==0.41.0
numpy==1.16.2
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
prettytable==0.7.2
prometheus-client==0.4.2
prompt-toolkit==2.0.7
ptyprocess==0.6.0
pycparser==2.19
pycurl==7.43.0
Pygments==2.2.0
pyparsing==2.2.0
pysam==0.15.1
python-dateutil==2.8.0
pytz==2018.9
PyVCF==0.6.8
pyzmq==17.1.2
qtconsole==4.4.2
recaptcha-client==1.0.6
s3transfer==0.2.0
scipy==1.2.1
Send2Trash==1.5.0
six==1.12.0
SQLAlchemy==1.2.12
sqlparse==0.2.4
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.2
urllib3==1.24.1
vine==1.2.0
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
=======
amqp==2.4.2
argon2-cffi==19.1.0
billiard==3.5.0.5
celery==4.2.1
cffi==1.12.2
confusable-homoglyphs==3.2.0
Django==2.1.7
django-crispy-forms==1.7.2
django-ranged-response==0.2.0
django-recaptcha==2.0.2
django-registration==3.0
kombu==4.4.0
mysqlclient==1.4.2.post1
numpy==1.16.2
Pillow==5.4.1
pycparser==2.19
pytz==2018.9
redis==3.2.0
scipy==1.2.1
six==1.12.0
urllib3==1.24.1
vine==1.2.0
>>>>>>> f54b90092c6a6fb9f26da8faad07ddf5460a4828
